flutter-game-v8: keep terrain layer indices strongly typed

+++ lib/core/formats.dart
@@ -976,8 +976,8 @@ class WorldData {
   int layerIndexAt(double x,double z){
     if(size==0||types.isEmpty||layers.isEmpty)return 0;
     final width=size~/2+1;
-    final ix=(x/2).floor().clamp(0,width-1);
-    final iz=(z/2).floor().clamp(0,width-1);
+    final ix=(x/2).floor().clamp(0,width-1).toInt();
+    final iz=(z/2).floor().clamp(0,width-1).toInt();
     final raw=types[iz*width+ix];
     return raw<layers.length?raw:0;
   }
